requirements.txt
openai
python-dotenv
redis
fastapi
uvicorn
